doubles correctly changed to floats

+++ gait/gait.h
@@ -11,43 +11,43 @@ typedef struct
 {
 	uint8_t type;
 
-	double period;
-	double periodstep;
-	double position;
+	float period;
+	float periodstep;
+	float position;
 	
-	double step_to_move_ratio;
-	double move_to_step_ratio;
-	double step_time;
-	double move_time;
+	float step_to_move_ratio;
+	float move_to_step_ratio;
+	float step_time;
+	float move_time;
 
-	double step_period_x;
-	double step_period_y;
-	double step_period_z;
+	float step_period_x;
+	float step_period_y;
+	float step_period_z;
 
-	double step_periodshift_x;
-	double step_periodshift_y;
-	double step_periodshift_z;
+	float step_periodshift_x;
+	float step_periodshift_y;
+	float step_periodshift_z;
 
-	double move_period_x;
-	double move_period_y;
-	double move_period_z;
+	float move_period_x;
+	float move_period_y;
+	float move_period_z;
 
-	double move_periodshift_x;
-	double move_periodshift_y;
-	double move_periodshift_z;
+	float move_periodshift_x;
+	float move_periodshift_y;
+	float move_periodshift_z;
 
-	double step_start[4];
-	double step_end[4];
+	float step_start[4];
+	float step_end[4];
 	
-	double x[4];
-	double y[4];
-	double z[4];
+	float x[4];
+	float y[4];
+	float z[4];
 } gaitdata;
 				
 void gait_process(gaitdata *g);
 void gait_paramcalc(gaitdata *g);
 void gait_init(gaitdata *g, uint8_t type);
-double gait_sine(double position, double Period, double period_shift, double amplitude, double amplitude_shift);
-double gait_line(double position, double Period, double period_shift, double amplitude, double amplitude_shift);
+float gait_sine(float position, float Period, float period_shift, float amplitude, float amplitude_shift);
+float gait_line(float position, float Period, float period_shift, float amplitude, float amplitude_shift);
 
 #endif
